lang/keywords: history stack for cd
This commit introduces a not persistent stack containing the last eight
directories the user navigated too via 'cd'. It also contains
improvements to feedback bubbled up to the user by introducing
consistent prefixes such as 'cd: '. As common with bash and other shells
'cd -' navigates to the last directory, popping a directory of the
aforementioned stack is done via 'cd ^', thus moving the cwd to the
previously visited directory and removing it from the stack.